improved project structure, added .gitignore and __init__.py and index.md
.gitignore is Python.gitignore from https://github.com/github/gitignore/blob/main/Python.gitignore
