+++ _src/generate_novelties.py
